pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharmarl"
version = "0.1.0"
description = "PharmaRL — first OpenEnv-native LLM-as-policy environment for iterative molecular optimization (Stage 1: DRD2 binding)"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "PyTDC>=0.4.1",
    "selfies>=2.1.1",
    "rdkit>=2023.9.5",
    "numpy>=1.24",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.27.0",
]
training = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "trl>=0.11.0",
    "unsloth>=2024.10",
    "wandb>=0.16.0",
    "datasets>=2.18.0",
]

[project.scripts]
server = "pharmarl.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["pharmarl", "pharmarl.server", "pharmarl.server.molecule_engine", "pharmarl.server.oracles"]
package-dir = { "pharmarl" = ".", "pharmarl.server" = "server", "pharmarl.server.molecule_engine" = "server/molecule_engine", "pharmarl.server.oracles" = "server/oracles" }
